+++ lib/src/imdb/db.rs
@@ -1,13 +1,8 @@
 #![warn(clippy::all)]
 
-use std::io::{BufRead, Write};
-
 use crate::imdb::db_impl::DbImpl;
-use crate::imdb::ratings::Ratings;
 use crate::imdb::title::Title;
-use crate::imdb::title::TsvAction;
 use crate::imdb::title_id::TitleId;
-use crate::utils::result::Res;
 use crate::utils::search::SearchString;
 
 use derive_more::{Display, From, Into};
@@ -23,64 +18,6 @@ pub enum Query {
   /// Query the database of Series.
   #[display(fmt = "series")]
   Series,
-}
-
-pub struct ServiceDb;
-
-impl ServiceDb {
-  /// Import title data from tab separated values (TSVs).
-  ///
-  /// This parses TSV data from the provided `ratings_reader` and `basics_reader` and
-  /// write them out in binary to the provided writers `movies_db_writer` and
-  /// `series_db_writer`.
-  ///
-  /// # Arguments
-  ///
-  /// * `ratings_reader` - TSV reader for ratings.
-  /// * `basics_reader` - TSV reader for title data.
-  /// * `movies_db_writer` - Binary writer to store movies.
-  /// * `series_db_writer` - Binary writer to store series.
-  pub(crate) fn import<R1: BufRead, R2: BufRead, W1: Write, W2: Write>(
-    ratings_reader: R1,
-    mut basics_reader: R2,
-    mut movies_db_writer: W1,
-    mut series_db_writer: W2,
-  ) -> Res<()> {
-    let ratings = Ratings::from_tsv(ratings_reader)?;
-
-    let mut line = String::new();
-
-    // Skip the first line.
-    basics_reader.read_line(&mut line)?;
-    line.clear();
-
-    loop {
-      let bytes = basics_reader.read_line(&mut line)?;
-
-      if bytes == 0 {
-        break;
-      }
-
-      let trimmed = line.trim_end();
-
-      if trimmed.is_empty() {
-        continue;
-      }
-
-      match Title::from_tsv(trimmed.as_bytes(), &ratings)? {
-        TsvAction::Movie(title) => title.write_binary(&mut movies_db_writer)?,
-        TsvAction::Series(title) => title.write_binary(&mut series_db_writer)?,
-        TsvAction::Skip => {
-          line.clear();
-          continue;
-        }
-      }
-
-      line.clear();
-    }
-
-    Ok(())
-  }
 }
 
 /// A special object (i.e. a handle) that is used to refer to a movie in the database.
@@ -233,10 +170,10 @@ impl Db {
 mod test_db {
   use std::io::Read;
 
-  use crate::imdb::db::ServiceDb;
   use crate::imdb::ratings::Ratings;
   use crate::imdb::testdata::{make_basics_reader, make_ratings_reader};
   use crate::imdb::title::Title;
+  use crate::imdb::tsv_import::tsv_import;
 
   #[test]
   fn test_to_binary() {
@@ -245,7 +182,7 @@ mod test_db {
 
     let mut movies_storage = Vec::new();
     let mut series_storage = Vec::new();
-    ServiceDb::import(ratings_reader, basics_reader, &mut movies_storage, &mut series_storage).unwrap();
+    tsv_import(ratings_reader, basics_reader, &mut movies_storage, &mut series_storage).unwrap();
 
     let mut basics_reader = make_basics_reader();
     let ratings_reader = make_ratings_reader();
